Extract run_cmd into utils.py to decouple CLI runner from pytest internals
